Refactor dataset and model backbone code for improved clarity and functionality
- Simplified the calculation of extra indices in `build_caries_oversampled_indices` by removing unnecessary rounding.
- Updated `_LayerFusion` to enforce strict mapping during tensor stacking, enhancing error handling.
- Adjusted `DinoV3Wrapper` to directly access `hidden_size` from the model configuration.
- Modified unit tests to reflect changes in tensor stacking and configuration handling, ensuring consistency across tests.

+++ src/dais26_dentex/data/dataset.py
@@ -111,7 +111,7 @@ def build_caries_oversampled_indices(
     expanded = list(base)
     for _ in range(full - 1):  # `base` already contributes one copy
         expanded.extend(positives)
-    extra = int(round(frac * len(positives)))
+    extra = round(frac * len(positives))
     expanded.extend(positives[:extra])
     return expanded
